move logical window function type-coercion logic from physical-expr to expr
