+++ datafusion/core/src/datasource/listing/helpers.rs
@@ -41,12 +41,12 @@ use crate::{
     error::Result,
     execution::context::SessionContext,
     logical_plan::{self, Expr, ExprVisitable, ExpressionVisitor, Recursion},
-    physical_plan::functions::Volatility,
     scalar::ScalarValue,
 };
 
 use super::{PartitionedFile, PartitionedFileStream};
 use datafusion_data_access::{object_store::ObjectStore, FileMeta, SizedFile};
+use datafusion_expr::Volatility;
 
 const FILE_SIZE_COLUMN_NAME: &str = "_df_part_file_size_";
 const FILE_PATH_COLUMN_NAME: &str = "_df_part_file_path_";
