+++ datafusion/core/src/logical_plan/expr_schema.rs
@@ -16,13 +16,11 @@
 // under the License.
 
 use super::Expr;
-use crate::physical_plan::{
-    expressions::binary_operator_data_type, functions, window_functions,
-};
+use crate::physical_plan::{expressions::binary_operator_data_type, functions};
 use arrow::compute::can_cast_types;
 use arrow::datatypes::DataType;
 use datafusion_common::{DFField, DFSchema, DataFusionError, ExprSchema, Result};
-use datafusion_expr::aggregate_function;
+use datafusion_expr::{aggregate_function, window_function};
 use datafusion_physical_expr::field_util::get_indexed_field;
 
 /// trait to allow expr to typable with respect to a schema
@@ -84,7 +82,7 @@ impl ExprSchemable for Expr {
                     .iter()
                     .map(|e| e.get_type(schema))
                     .collect::<Result<Vec<_>>>()?;
-                window_functions::return_type(fun, &data_types)
+                window_function::return_type(fun, &data_types)
             }
             Expr::AggregateFunction { fun, args, .. } => {
                 let data_types = args
