+++ datafusion/core/src/sql/planner.rs
@@ -40,14 +40,13 @@ use crate::scalar::ScalarValue;
 use crate::sql::utils::{make_decimal_type, normalize_ident};
 use crate::{
     error::{DataFusionError, Result},
+    physical_plan::aggregates,
     physical_plan::udaf::AggregateUDF,
-};
-use crate::{
     physical_plan::udf::ScalarUDF,
-    physical_plan::{aggregates, functions, window_functions},
     sql::parser::{CreateExternalTable, FileType, Statement as DFStatement},
 };
 use arrow::datatypes::*;
+use datafusion_expr::{window_function::WindowFunction, BuiltinScalarFunction};
 use hashbrown::HashMap;
 
 use sqlparser::ast::{
@@ -1417,7 +1416,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             SQLExpr::Value(Value::Boolean(n)) => Ok(lit(n)),
             SQLExpr::Value(Value::Null) => Ok(Expr::Literal(ScalarValue::Utf8(None))),
             SQLExpr::Extract { field, expr } => Ok(Expr::ScalarFunction {
-                fun: functions::BuiltinScalarFunction::DatePart,
+                fun: BuiltinScalarFunction::DatePart,
                 args: vec![
                     Expr::Literal(ScalarValue::Utf8(Some(format!("{}", field)))),
                     self.sql_expr_to_logical_expr(*expr, schema)?,
@@ -1671,7 +1670,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
 
 
                 Ok(Expr::ScalarFunction {
-                    fun: functions::BuiltinScalarFunction::Substr,
+                    fun: BuiltinScalarFunction::Substr,
                     args,
                 })
             }
@@ -1688,15 +1687,15 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
             SQLExpr::Trim { expr, trim_where } => {
                 let (fun, where_expr) = match trim_where {
                     Some((TrimWhereField::Leading, expr)) => {
-                        (functions::BuiltinScalarFunction::Ltrim, Some(expr))
+                        (BuiltinScalarFunction::Ltrim, Some(expr))
                     }
                     Some((TrimWhereField::Trailing, expr)) => {
-                        (functions::BuiltinScalarFunction::Rtrim, Some(expr))
+                        (BuiltinScalarFunction::Rtrim, Some(expr))
                     }
                     Some((TrimWhereField::Both, expr)) => {
-                        (functions::BuiltinScalarFunction::Btrim, Some(expr))
+                        (BuiltinScalarFunction::Btrim, Some(expr))
                     }
-                    None => (functions::BuiltinScalarFunction::Trim, None),
+                    None => (BuiltinScalarFunction::Trim, None),
                 };
                 let arg = self.sql_expr_to_logical_expr(*expr, schema)?;
                 let args = match where_expr {
@@ -1719,7 +1718,7 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                 };
 
                 // first, scalar built-in
-                if let Ok(fun) = functions::BuiltinScalarFunction::from_str(&name) {
+                if let Ok(fun) = BuiltinScalarFunction::from_str(&name) {
                     let args = self.function_args_to_expr(function.args, schema)?;
 
                     return Ok(Expr::ScalarFunction { fun, args });
@@ -1752,13 +1751,13 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                             }
                         })
                         .transpose()?;
-                    let fun = window_functions::WindowFunction::from_str(&name)?;
+                    let fun = WindowFunction::from_str(&name)?;
                     match fun {
-                        window_functions::WindowFunction::AggregateFunction(
+                        WindowFunction::AggregateFunction(
                             aggregate_fun,
                         ) => {
                             return Ok(Expr::WindowFunction {
-                                fun: window_functions::WindowFunction::AggregateFunction(
+                                fun: WindowFunction::AggregateFunction(
                                     aggregate_fun.clone(),
                                 ),
                                 args: self.aggregate_fn_to_expr(
@@ -1771,11 +1770,11 @@ impl<'a, S: ContextProvider> SqlToRel<'a, S> {
                                 window_frame,
                             });
                         }
-                        window_functions::WindowFunction::BuiltInWindowFunction(
+                        WindowFunction::BuiltInWindowFunction(
                             window_fun,
                         ) => {
                             return Ok(Expr::WindowFunction {
-                                fun: window_functions::WindowFunction::BuiltInWindowFunction(
+                                fun: WindowFunction::BuiltInWindowFunction(
                                     window_fun,
                                 ),
                                 args: self.function_args_to_expr(function.args, schema)?,
@@ -2307,9 +2306,8 @@ fn parse_sql_number(n: &str) -> Result<Expr> {
 #[cfg(test)]
 mod tests {
     use crate::datasource::empty::EmptyTable;
-    use crate::physical_plan::functions::Volatility;
     use crate::{assert_contains, logical_plan::create_udf, sql::parser::DFParser};
-    use datafusion_expr::ScalarFunctionImplementation;
+    use datafusion_expr::{ScalarFunctionImplementation, Volatility};
 
     use super::*;
